# api/docs/swagger.yaml
basePath: /
definitions:
  budgeting_service.Account:
    properties:
      balance:
        type: number
      created_at:
        type: string
      currency:
        type: string
      id:
        type: string
      name:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  budgeting_service.Accounts:
    properties:
      accounts:
        items:
          $ref: '#/definitions/budgeting_service.Account'
        type: array
    type: object
  budgeting_service.Budget:
    properties:
      amount:
        type: number
      category_id:
        type: string
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      period:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  budgeting_service.Budgets:
    properties:
      budgets:
        items:
          $ref: '#/definitions/budgeting_service.Budget'
        type: array
    type: object
...
  budgeting_service.CreateAccount:
    properties:
      balance:
        type: number
      currency:
        type: string
      name:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  budgeting_service.CreateBudget:
    properties:
      amount:
        type: number
      category_id:
        type: string
      end_date:
        type: string
      period:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
...
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
...
info:
  contact: {}
  description: Something big
  title: MoneyMate
  version: "1.0"
paths:
  /accounts/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves an account by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account retrieved successfully
          schema:
            $ref: '#/definitions/budgeting_service.Account'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      tags:
      - accounts
  /accounts/{id}/delete:
    delete:
      consumes:
      - application/json
      description: Deletes an account by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account deleted successfully
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      tags:
      - accounts
  /accounts/{id}/update:
    put:
      consumes:
      - application/json
      description: Updates an existing account
      parameters:
      - description: Account id
        in: path
        name: id
        required: true
        type: string
      - description: Account Update Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/budgeting_service.Account'
      produces:
      - application/json
      responses:
        "200":
          description: Account updated successfully
          schema:
            $ref: '#/definitions/budgeting_service.Account'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      tags:
      - accounts
  /accounts/all:
    get:
      consumes:
      - application/json
      description: Retrieves all accounts
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: Limit number
        in: query
        name: limit
        required: true
        type: integer
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Name
        in: query
        name: name
        type: string
      - description: Type
        in: query
        name: type
        type: string
      - description: Balance from
        in: query
        name: balance_from
        type: number
      - description: Balance to
        in: query
        name: balance_to
        type: number
      - description: Currency
        in: query
        name: currency
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Accounts retrieved successfully
          schema:
            $ref: '#/definitions/budgeting_service.Accounts'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      tags:
      - accounts
  /accounts/create:
    post:
      consumes:
      - application/json
      description: Creates an account
      parameters:
      - description: Account Creation Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/budgeting_service.CreateAccount'
      produces:
      - application/json
      responses:
        "201":
          description: Account created successfully
          schema:
            $ref: '#/definitions/budgeting_service.Account'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      tags:
      - accounts
  /budgets/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a budget by ID
      parameters:
      - description: Budget ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Budget retrieved successfully
          schema:
            $ref: '#/definitions/budgeting_service.Budget'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      tags:
      - budgets
  /budgets/{id}/delete:
    delete:
      consumes:
      - application/json
      description: Deletes a budget by ID
      parameters:
      - description: Budget ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Budget deleted successfully
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      tags:
      - budgets
  /budgets/{id}/update:
    put:
      consumes:
      - application/json
      description: Updates an existing budget
      parameters:
      - description: Budget ID
        in: path
        name: id
        required: true
        type: string
      - description: Budget Update Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/budgeting_service.Budget'
      produces:
      - application/json
      responses:
        "200":
          description: Budget updated successfully
          schema:
            $ref: '#/definitions/budgeting_service.Budget'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      tags:
      - budgets
  /budgets/all:
    get:
      consumes:
      - application/json
      description: Retrieves all budgets
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: Limit number
        in: query
        name: limit
        required: true
        type: integer
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: category_id
        in: query
        name: category_id
        type: string
      - description: Name
        in: query
        name: period
        type: string
      - description: Type
        in: query
        name: type
        type: string
      - description: Amount
        in: query
        name: amount
        type: number
      - description: start_date
        in: query
        name: start_date
        type: string
      - description: end_date
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Budgets retrieved successfully
          schema:
            $ref: '#/definitions/budgeting_service.Budgets'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      tags:
      - budgets
  /budgets/create:
    post:
      consumes:
      - application/json
      description: Creates a new budget
      parameters:
      - description: Budget Creation Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/budgeting_service.CreateBudget'
      produces:
      - application/json
      responses:
        "201":
          description: Budget created successfully
          schema:
            $ref: '#/definitions/budgeting_service.Budget'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      tags:
      - budgets
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"